refactor #109 OpinionType 공통모듈화
- feature/chat에서만 쓰던 enum을 여러 feature에서 사용하게 되어 공통화 진행
- `/core/enums` 디렉터리 생성
