+++ lib/core/enums/opinion_type.dart
